+++ packages/utils/theme/src/types.ts
@@ -169,8 +169,7 @@ export interface Theme {
     onOverlay: string
   }
   fontFamily: {
-    nunitoSans: string
-    openSans: string
+    sans: string
   }
   /**
    * https://www.figma.com/file/ezqkDE3LqYrek9MvYnr4dR/Spark-Foundations?node-id=105%3A2710&t=lLeJnznf64zUBmdL-4