+++ packages/components/form-field/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1](https://github.com/adevinta/spark/compare/@spark-ui/form-field@0.1.0...@spark-ui/form-field@0.1.1) (2023-05-22)
+
+### Bug Fixes
+
+- **form-field:** update missing provider error message ([6936263](https://github.com/adevinta/spark/commit/6936263b0ba08dc2c1978841877543eea710db2a))
+
 # 0.1.0 (2023-05-12)
 
 ### Features