+++ packages/components/icon/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.2](https://github.com/adevinta/spark/compare/@spark-ui/icon@1.7.1...@spark-ui/icon@1.7.2) (2023-05-24)
+
+### Bug Fixes
+
+- **icon:** change children type to be a node ([42003f7](https://github.com/adevinta/spark/commit/42003f7fb76b3de90c5762f38e2d0f1b5e290c29))
+
+## [1.7.1](https://github.com/adevinta/spark/compare/@spark-ui/icon@1.7.0...@spark-ui/icon@1.7.1) (2023-05-22)
+
+**Note:** Version bump only for package @spark-ui/icon
+
 # [1.7.0](https://github.com/adevinta/spark/compare/@spark-ui/icon@1.6.5...@spark-ui/icon@1.7.0) (2023-05-11)
 
 ### Features