+++ packages/components/switch/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.17](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.10.16...@spark-ui/switch@1.10.17) (2023-05-24)
+
+**Note:** Version bump only for package @spark-ui/switch
+
 ## [1.10.16](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.10.15...@spark-ui/switch@1.10.16) (2023-05-17)
 
 **Note:** Version bump only for package @spark-ui/switch